pyproject.toml
[project]
name = "frontend-geo-mcp"
version = "0.1.0"
description = "MCP Frontend GPS - Navigator and Code Reviewer for React projects"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.13.0",
    "supabase>=2.0.0",
    "postgrest>=0.16.0",
    "gitpython>=3.1.0",
    "aiofiles>=23.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "typing-extensions>=4.9.0",
    "psycopg2-binary>=2.9.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[[tool.hatch.build.targets.wheel.force-include]]
path = "src"
destination = "src"
